requirements.txt
alembic==1.5.7
bcrypt==3.2.0
boto3==1.17.29
GitPython==3.1.14
Flask==1.1.2
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-Limiter==1.4
Flask-Migrate==2.7.0
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
password-strength==0.0.3.post2
psycopg2==2.8.6
PyJWT==2.0.1
pyotp==2.6.0
redis==3.5.3